requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.5.3
httpx==0.26.0
pandas==2.2.0
openpyxl==3.1.2
playwright==1.41.0
supabase==2.3.4
python-dotenv==1.0.0
